refactor(headless-client): remove auto token detection
